Still trying to make heroku happy:
Gem previously only developed using local sqlite db, but the migration
block where we specify a type as :string breaks when the migration
attempts to run on postgres (which heroku use it seems).

Switched the migration to use the class name String instead of the
symbol as the postgres driver seems happier with that.

Also removed the dependency on capistrano, deploying to heroko obviates
the need for capistrano (at least as long as it's free).