Previous commit was cute and all, but still needed the settings file to exist.
Removed that completely, as it doesn't on heroku. Just go with the
heroku flow, and make the config settings just load from env variables.
A couple of cases it was possible to have meaningful defaults, but added
a hard exception if the basic auth username and password aren't set in
the appropriate env variables.